+++ src/index.ts
@@ -7,12 +7,18 @@ export {
   createSessionStorage,
   createIndexedDBStorage,
   persist,
-  persistBrowserLocal,
-  persistBrowserSession,
-  persistCookie,
   disableWarnings,
   setSerialization,
 } from "./core"
+export {
+  persistBrowserLocal,
+  persistBrowserSession,
+  persistCookie,
+  localWritable,
+  writable,
+  sessionWritable,
+  cookieWritable,
+} from "./alias"
 export type { CHROME_STORAGE_TYPE, PersistentStore, StorageInterface, SelfUpdateStorageInterface } from "./core"
 export { createEncryptionStorage, createEncryptedStorage, noEncryptionBehavior } from "./encryption"
 export type { NO_ENCRYPTION_BEHAVIOR, Encryption, GCMEncryption } from "./encryption"